Rename files
To keep it consistent with main, I have renamed files that are not part
of app from filename to "_filename" out of app. I also cleaned up the
outdated user-service because it was renamed already.
