+++ frontend/src/_helpers/interceptors/jwt.interceptor.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@angular/core';
 import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
-import { environment } from '../../environments/environment';
-import { AuthenticationService } from '../../services/authentication.service';
+import { environment } from '../../_environments/environment';
+import { AuthenticationService } from '../../_services/authentication.service';
 
 @Injectable()
 export class JwtInterceptor implements HttpInterceptor {
