+++ frontend/src/app/app.config.ts
@@ -3,8 +3,8 @@ import { provideRouter } from '@angular/router';
 import { HTTP_INTERCEPTORS, provideHttpClient } from '@angular/common/http';
 import { routes } from './app.routes';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
-import { JwtInterceptor } from './helpers/interceptors/jwt.interceptor';
-import { ErrorInterceptor } from './helpers/interceptors/error.interceptor';
+import { JwtInterceptor } from '../_helpers/interceptors/jwt.interceptor';
+import { ErrorInterceptor } from '../_helpers/interceptors/error.interceptor';
 
 export const appConfig: ApplicationConfig = {
     providers: [
