--- services/user-service/.gitignore
@@ -1,7 +1,0 @@
-**/node_modules
-**/.env
-**/.idea
-# Vim temp files
-*~
-*.swp
-*.swo
